+++ backend/app/main.py
@@ -3,8 +3,8 @@ from fastapi import Depends, FastAPI, File, Header, HTTPException, UploadFile
 from fastapi_limiter.depends import RateLimiter
 from google import genai
 
-from parse_bill import parse_bill
-from rate_limiting import lifespan
+from app.utils.parse_bill import parse_bill
+from app.utils.rate_limiting import lifespan
 
 app = FastAPI(lifespan=lifespan)
 
@@ -16,10 +16,13 @@ GEMINI_API_KEY = config("GEMINI_API_KEY")
 GEMINI_MODEL_NAME = config("GEMINI_MODEL_NAME", default="gemini-1.5-flash")
 
 
-@app.get("/", dependencies=[
-    Depends(RateLimiter(times=3, minutes=1)),
-    Depends(RateLimiter(times=10, hours=1)),
-])
+@app.get(
+    "/",
+    dependencies=[
+        Depends(RateLimiter(times=3, minutes=1)),
+        Depends(RateLimiter(times=10, hours=1)),
+    ],
+)
 def index():
     return {"Hello": "World"}
 
@@ -49,10 +52,13 @@ def generate_gemini(prompt: str, x_api_key: str = Depends(verify_api_key)):
         raise HTTPException(status_code=500, detail=f"Gemini API error: {e}")
 
 
-@app.post("/parse_bill", dependencies=[
-    Depends(RateLimiter(times=1, minutes=1)),
-    Depends(RateLimiter(times=10, hours=1)),
-])
+@app.post(
+    "/parse_bill",
+    dependencies=[
+        Depends(RateLimiter(times=1, minutes=1)),
+        Depends(RateLimiter(times=10, hours=1)),
+    ],
+)
 async def upload_bill(
     file: UploadFile = File(...), x_api_key: str = Depends(verify_api_key)
 ):
@@ -60,7 +66,7 @@ async def upload_bill(
         raise HTTPException(
             status_code=400, detail="Invalid file type. Please upload an image."
         )
-    
+
     API_KEY_CREDITS[x_api_key] -= 1
     return await parse_bill(file)
